project: major update - rm'd registering, rm'd profile, bugfixes
It is no longer possible for the user to register new users. Instead, a default admin user is created when the server is first started. Client may change the login credentials of the default admin user by logging in to the app and navigating to change credentials page.

Profile has also been removed because it is unnecessary.

Few bugfixes and user experience improvements have been implemented.